feat(be): move jwt-auth into auth/jwt lib
