+++ backend/.swcrc
@@ -15,6 +15,7 @@
       "@client/*": ["./apps/client/src/*"],
       "@libs/prisma": ["./libs/prisma/src/index.ts"],
       "@libs/cache": ["./libs/cache/src/index.ts"],
+      "@libs/auth": ["libs/auth/src/index.ts"],
       "@libs/decorator": ["libs/decorator/src/index.ts"],
       "@libs/exception": ["libs/exception/src/index.ts"],
       "@libs/pipe": ["libs/pipe/src/index.ts"],
