+++ backend/apps/admin/src/admin.module.ts
@@ -3,9 +3,8 @@ import { Module } from '@nestjs/common'
 import { ConfigModule } from '@nestjs/config'
 import { APP_GUARD } from '@nestjs/core'
 import { GraphQLModule } from '@nestjs/graphql'
+import { JwtAuthGuard, JwtAuthModule } from '@libs/auth'
 import { PrismaModule } from '@libs/prisma'
-import { AuthModule } from '@admin/auth/auth.module'
-import { JwtAuthGuard } from '@admin/auth/guard/jwt-auth.guard'
 import { AdminController } from './admin.controller'
 import { AdminService } from './admin.service'
 import { RolesGuard } from './user/guard/roles.guard'
@@ -19,9 +18,9 @@ import { UserModule } from './user/user.module'
       autoSchemaFile: 'schema.gql',
       sortSchema: true
     }),
+    JwtAuthModule,
     UserModule,
-    PrismaModule,
-    AuthModule
+    PrismaModule
   ],
   controllers: [AdminController],
   providers: [
