--- backend/apps/admin/src/auth/auth.module.ts
@@ -1,27 +1,0 @@
-import { Module } from '@nestjs/common'
-import { ConfigModule, ConfigService } from '@nestjs/config'
-import { JwtModule, type JwtModuleOptions } from '@nestjs/jwt'
-import { PassportModule } from '@nestjs/passport'
-import { JwtStrategy } from './strategy/jwt.strategy'
-
-@Module({
-  imports: [
-    PassportModule,
-    JwtModule.registerAsync({
-      imports: [ConfigModule],
-      useFactory: async (config: ConfigService) => {
-        const options: JwtModuleOptions = {
-          secret: config.get('JWT_SECRET'),
-          signOptions: {
-            expiresIn: '10m',
-            issuer: 'skkuding.dev'
-          }
-        }
-        return options
-      },
-      inject: [ConfigService]
-    })
-  ],
-  providers: [JwtStrategy]
-})
-export class AuthModule {}
