--- backend/apps/admin/src/auth/guard/jwt-auth.guard.ts
@@ -1,22 +1,0 @@
-import { type ExecutionContext, Injectable } from '@nestjs/common'
-import { Reflector } from '@nestjs/core'
-import { GqlExecutionContext } from '@nestjs/graphql'
-import { AuthGuard } from '@nestjs/passport'
-import type { Observable } from 'rxjs'
-
-@Injectable()
-export class JwtAuthGuard extends AuthGuard('jwt') {
-  constructor(private reflector: Reflector) {
-    super()
-  }
-
-  canActivate(
-    context: ExecutionContext
-  ): boolean | Promise<boolean> | Observable<boolean> {
-    return super.canActivate(context)
-  }
-
-  getRequest(context: ExecutionContext) {
-    return GqlExecutionContext.create(context).getContext().req
-  }
-}
