+++ backend/apps/client/src/app.module.ts
@@ -3,12 +3,12 @@ import { CacheModule } from '@nestjs/cache-manager'
 import { Module } from '@nestjs/common'
 import { ConfigModule } from '@nestjs/config'
 import { APP_GUARD } from '@nestjs/core'
+import { JwtAuthGuard, JwtAuthModule } from '@libs/auth'
 import { CacheConfigService } from '@libs/cache'
 import { PrismaModule } from '@libs/prisma'
 import { AppController } from './app.controller'
 import { AppService } from './app.service'
 import { AuthModule } from './auth/auth.module'
-import { JwtAuthGuard } from './auth/guard/jwt-auth.guard'
 import { ClarificationModule } from './clarification/clarification.module'
 import { ContestModule } from './contest/contest.module'
 import { EmailModule } from './email/email.module'
@@ -27,6 +27,7 @@ import { WorkbookModule } from './workbook/workbook.module'
       isGlobal: true,
       useClass: CacheConfigService
     }),
+    JwtAuthModule,
     MailerModule.forRootAsync({
       useClass: MailerConfigService
     }),
