+++ backend/apps/client/src/auth/auth.controller.ts
@@ -9,13 +9,13 @@ import {
   InternalServerErrorException
 } from '@nestjs/common'
 import { Request, Response } from 'express'
+import type { JwtTokens } from '@libs/auth'
 import { REFRESH_TOKEN_COOKIE_OPTIONS, AUTH_TYPE } from '@libs/constants'
 import { AuthNotNeeded } from '@libs/decorator'
 import { InvalidJwtTokenException, InvalidUserException } from '@libs/exception'
 import { AuthService } from './auth.service'
 import { LoginUserDto } from './dto/login-user.dto'
 import { AuthenticatedRequest } from './interface/authenticated-request.interface'
-import type { JwtTokens } from './interface/jwt.interface'
 
 @Controller('auth')
 export class AuthController {
