+++ backend/apps/client/src/auth/interface/authenticated-request.interface.ts
@@ -1,5 +1,5 @@
 import type { Request } from 'express'
-import type { AuthenticatedUser } from '../class/authenticated-user.class'
+import type { AuthenticatedUser } from '@libs/auth'
 
 export interface AuthenticatedRequest extends Request {
   user: AuthenticatedUser
