+++ backend/libs/auth/src/authenticated-user.class.ts
