+++ backend/libs/auth/src/index.ts
@@ -1,0 +1,5 @@
+export * from './authenticated-user.class'
+
+export * from './jwt/jwt-auth.module'
+export * from './jwt/jwt-auth.guard'
+export * from './jwt/jwt.interface'
