+++ backend/libs/auth/src/jwt/jwt-auth.guard.ts
@@ -1,0 +1,35 @@
+import { type ExecutionContext, Injectable } from '@nestjs/common'
+import { Reflector } from '@nestjs/core'
+import { type GqlContextType, GqlExecutionContext } from '@nestjs/graphql'
+import { AuthGuard } from '@nestjs/passport'
+import type { Observable } from 'rxjs'
+import { IS_AUTH_NOT_NEEDED_KEY } from '@libs/decorator'
+
+@Injectable()
+export class JwtAuthGuard extends AuthGuard('jwt') {
+  constructor(private reflector: Reflector) {
+    super()
+  }
+
+  canActivate(
+    context: ExecutionContext
+  ): boolean | Promise<boolean> | Observable<boolean> {
+    if (context.getType<GqlContextType>() !== 'graphql') {
+      const isAuthNotNeeded = this.reflector.getAllAndOverride<boolean>(
+        IS_AUTH_NOT_NEEDED_KEY,
+        [context.getHandler(), context.getClass()]
+      )
+      if (isAuthNotNeeded) {
+        return true
+      }
+    }
+    return super.canActivate(context)
+  }
+
+  getRequest(context: ExecutionContext) {
+    if (context.getType<GqlContextType>() === 'graphql') {
+      return GqlExecutionContext.create(context).getContext().req
+    }
+    return super.getRequest(context)
+  }
+}
