+++ backend/libs/auth/src/jwt/jwt-auth.module.ts
@@ -1,0 +1,9 @@
+import { Module } from '@nestjs/common'
+import { PassportModule } from '@nestjs/passport'
+import { JwtStrategy } from './jwt.strategy'
+
+@Module({
+  imports: [PassportModule],
+  providers: [JwtStrategy]
+})
+export class JwtAuthModule {}
