+++ backend/libs/auth/src/jwt/jwt.strategy.ts
@@ -3,8 +3,8 @@ import { ConfigService } from '@nestjs/config'
 import { PassportStrategy } from '@nestjs/passport'
 import { ExtractJwt } from 'passport-jwt'
 import { Strategy } from 'passport-jwt'
-import { AuthenticatedUser } from '../class/authenticated-user.class'
-import type { JwtObject } from '../interface/jwt.interface'
+import { AuthenticatedUser } from '../authenticated-user.class'
+import type { JwtObject } from './jwt.interface'
 
 @Injectable()
 export class JwtStrategy extends PassportStrategy(Strategy) {
