+++ backend/nest-cli.json
@@ -53,6 +53,15 @@
         "tsConfigPath": "libs/cache/tsconfig.lib.json"
       }
     },
+    "auth": {
+      "type": "library",
+      "root": "libs/auth",
+      "entryFile": "index",
+      "sourceRoot": "libs/auth/src",
+      "compilerOptions": {
+        "tsConfigPath": "libs/auth/tsconfig.lib.json"
+      }
+    },
     "decorator": {
       "type": "library",
       "root": "libs/decorator",
